requirements.txt
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
curl_cffi==0.13.0
filelock==3.20.1
frozendict==2.4.7
fsspec==2025.12.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multitasking==0.0.12
networkx==3.6.1
numpy==2.3.5
openbb==4.6.0
openbb-yfinance==1.5.2
packaging==25.0
pandas==2.3.3
peewee==3.18.3
platformdirs==4.5.1
protobuf==6.33.2
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.6.0
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
websockets==15.0.1
yfinance==0.2.66
